% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptors.R
\name{get_descriptors}
\alias{get_descriptors}
\title{Descriptors}
\usage{
get_descriptors(x, type = c("params", "service", "status", "copyright"))
}
\arguments{
\item{x}{An object returned by a \code{z22_*} function.}

\item{type}{The type of descriptor to return.}
}
\value{
A list of response metadata of a \code{z22_*} object.
}
\description{
Descriptors represent the metadata added to each Zensus response. They
include request parameters, service details, status information, and
copyright notices.
}
\examples{
\dontrun{
# get a z22 object
res <- z22_statistics()

# retrieve its descriptors
get_descriptors(res)
}
}
