% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z22-package.R
\docType{package}
\name{z22-package}
\alias{z22}
\alias{z22-package}
\title{z22: Official Gridded Data from the German Census 2022}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org} [copyright holder]

}
\keyword{internal}
