% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{z22_auth}
\alias{z22_auth}
\title{Authenticate}
\usage{
z22_auth(
  token = NULL,
  username = NULL,
  password = askpass::askpass,
  overwrite = FALSE
)
}
\arguments{
\item{token}{API token that can be retrieved from the
\href{https://ergebnisse2011.zensus2022.de/datenbank/online}{Zensus 2022 database}.}

\item{username}{Zensus2022 username.}

\item{password}{Function that prompts a password entry. Defaults to
\code{\link[askpass]{askpass}}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Stores Zensus2022 login credentials or tokens to access API methods that
require authentication. Login credentials are securely stored using
the \code{\link{keyring}} package and are automatically retrieved to
make authenticated requests.

\code{z22_auth} supports authentication using an an API token or
using credentials (i.e. username and password). If an API token is
provided, it is stored inside an R environment for further use.
If a username and a password function is provided, the password is
created by executing the password function. It is then stored in a keyring.
}
\note{
Authentication using username and password does not work currently.
}
\examples{
\dontrun{
# authenticate using a token
z22_auth(token = "XXXxxXXxxXX")

# verify that the authentication worked
z22_is_auth()

# authenticate using credentials
z22_auth(username = "user")

# using a different password prompt
z22_auth(username = "user", password = rstudioapi::askForPassword)
}
}
\seealso{
\code{\link{z22_is_auth}}
}
