% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{z22_catalogue}
\alias{z22_catalogue}
\title{Zensus catalogue}
\usage{
z22_catalogue(object, selection = NULL, ..., pagelength = 100L)
}
\arguments{
\item{object}{The type of object for which to return lists.}

\item{selection}{A search string with a maximum of 15 characters. The
statistics catalog is searched for the selection string. The
\code{searchcriterion} can be used to specify which column should be
searched. Search strings can use the wildcard operator \code{*}.}

\item{...}{Further arguments passed to \code{\link{z22_statistics}},
\code{\link{z22_variables}}, \code{\link{z22_tables}},
\code{\link{z22_values}}, or \code{\link{z22_results}}.}

\item{pagelength}{Number of results to return up to a maximum of 2,500.
Defaults to 100.}
}
\value{
A dataframe containing information on the code, content, and
number of variables.
}
\description{
Retrieve census contents of statistics, variables, tables, values,
and results.
}
