% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{z22_decode_feat}
\alias{z22_decode_feat}
\alias{z22_decode_cat}
\alias{z22_translate_feat}
\alias{z22_translate_cat}
\title{Decode and translate features and categories}
\usage{
z22_decode_feat(features, lang = c("english", "german"))

z22_decode_cat(features, codes = NULL, lang = c("english", "german"))

z22_translate_feat(
  features,
  type = c("desc", "name"),
  lang = c("english", "german")
)

z22_translate_cat(categories, lang = c("english", "german"))
}
\arguments{
\item{features}{A character vector containing English or German feature
names.}

\item{lang}{Specifies the language of the output description. Can be
either \code{"english"} (default) or \code{"german"}. Note that the
English descriptions are only ad-hoc translations based off the German
originals.}

\item{codes}{A numeric vector containing category codes. If \code{NULL},
returns all category descriptions within the feature specified in the
\code{feature} argument.}
}
\value{
A character vector containing the English or German descriptions
of the features or categories.

\code{z22_decode_cat} always returns a named list where each element is
named by its feature and contains the features matched category
descriptions. For consistency, this is also the case if \code{features}
only contains a single feature name.
}
\description{
Helper functions to get the English or German descriptions associated with
a feature name or category code.

\itemize{
 \item{The \code{z22_decode_*} functions help to retrieve human-readable
 descriptions from abbreviated feature names}
 \item{The \code{z22_translate_*} functions help to translate feature names
 or descriptions from German to English or vice versa}
}
}
\examples{
# Get english description of feature names
z22_decode_feat(c("ALTER_KURZ", "ALTER_10JG"))

# You can use both german and english feature names.
# English feature names are not official but might be better
# for reproducibility.
z22_decode_feat(c("age_short", "age_long"))

# Get all categories for specific features
z22_decode_cat(c("build_size", "build_heating"))

# Translate feature names
z22_translate_feat(
  c("build_size", "build_heating"),
  lang = "german",
  type = "name"
)

# Translate category labels
z22_translate_cat("Eingetr. Lebenspartnerschaft")
}
