% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{z22_find}
\alias{z22_find}
\title{Find an object}
\usage{
z22_find(term = NULL, category = NULL, pagelength = 100)
}
\arguments{
\item{term}{Search term.}

\item{category}{Category in which to search. Must be one of \code{"all"},
\code{"tables"}, \code{"statistics"}, or \code{"variables"}.}

\item{pagelength}{Number of results to return up to a maximum of 2,500.
Defaults to 100.}
}
\value{
A dataframe containing the search results in long format.
The \code{type} column specifies the type of search result. Every result
includes columns \code{Code} and \code{Content}. Depending on the result
type, the dataframe can contain additional columns.
}
\description{
Search for a table, statistic, or variable.
}
\examples{
\dontrun{
# search for table, statistics, and variables containing "population"
z22_find("population")

# restrict the search to statistics
z22_find("population", category = "statistics")
}
}
