% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes.R
\name{z22_get_attribute_100m}
\alias{z22_get_attribute_100m}
\alias{z22_get_attribute_1km}
\title{Get Census 2022 gridded attribute}
\usage{
z22_get_attribute_100m(
  topic,
  feature,
  category,
  all_cells = FALSE,
  rasterize = FALSE,
  as_sf = FALSE,
  update_cache = FALSE
)

z22_get_attribute_1km(
  feature,
  type = "ordinal",
  all_cells = TRUE,
  rasterize = FALSE,
  as_sf = FALSE,
  update_cache = FALSE
)
}
\arguments{
\item{topic, feature, category}{The topic, feature and category of the
request attribute. A list of available combinations can be retrieved
using \code{\link{z22_list_attributes}}. With the exception of
\code{topic = "population"} (where only the topic can be specified), all
arguments must be provided.}

\item{all_cells}{If \code{TRUE}, joins the retrieved attribute with the
complete grid from \code{\link{z22_grid}}. Otherwise, the attribute grid
will contain only those grid cells with one or more recorded units. Defaults
to \code{FALSE}, because loading the grid and joining with it is
computationally expensive.}

\item{rasterize}{If \code{TRUE} and the \code{terra} package is installed,
converts the attribute coordinates to a \code{\link[terra:rast]{SpatRaster}}.}

\item{as_sf}{If \code{TRUE} and the \code{sf} package is installed,
converts the attribute coordinates to an \code{sf} tibble.}

\item{update_cache}{By default, both functions cache attribute files for
the remainder of the R session. They are downloaded to a temporary directory
and - if the file to download already exists - are recovered from the cache.
In other words, when rerunning the same request multiple times, the
subsequent calls should be much faster. If \code{TRUE}, disables caching
for this call and overwrites the currently cached attribute file (if any)
with a fresh one. Defaults to \code{FALSE}, i.e. always cache.}

\item{res}{Resolution of the grid dataset. Can be \code{"100m"}
or \code{"1km"}.}
}
\value{
A tibble, \code{\link[terra:rast]{SpatRaster}} or
\code{\link[sf::st_as_sf]{sf}} tibble depending on the \code{rasterize}
and \code{as_sf} arguments.
}
\description{
Retrieve the values and coordinates of gridded attributes from the Census
2022.

An attribute is a thematic chunk divided by three aspects:

\itemize{
 \item{The general \strong{topic} of the attribute, e.g. demography or
 families (\code{topic})}
 \item{A single spatial characteristic or \strong{feature}, e.g. age or
 household size (\code{feature})}
 \item{The levels or \strong{categories} of each feature, e.g. single age
 groups (\code{category})}
}
}
\details{
Half of the grids cell width is added to each coordinate in the
dataset internally. According to the INSPIRE guidelines, coordinates
always represent the South-west of the grid cells. Centroids represent
the geographic location of grid cells better which is why they are used.
}
\examples{
\donttest{# Get gridded population
pop <- z22_get_attribute("population", all_cells = TRUE, rasterize = TRUE)
terra::plot(pop)

# Get data about buildings using district heating
z22_get_attribute("buildings", "HEIZTYP", 1)}
}
