% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{z22_is_auth}
\alias{z22_is_auth}
\title{Verify authentication}
\usage{
z22_is_auth(error = FALSE)
}
\arguments{
\item{error}{Whether to throw an error if authentication failed.
If \code{FALSE}, returns a logical value.}
}
\value{
A logical value indicating whether the authentication was
successful. If \code{error = TRUE}, throws an error if this value is
\code{FALSE}.
}
\description{
Verify authentication using \code{\link{z22_auth}} by querying the
\code{logincheck} method.
}
\examples{
\dontrun{
# authenticate
z22_auth(token = "XXXxxXXxxXX")

# verify that the authentication worked
z22_is_auth()

# use to gatekeep further analysis
# if authentication fails, aborts operation before
# other methods are requested
z22_is_auth(error = TRUE)
z22_find()
}
}
\seealso{
\code{\link{z22_auth}}
}
