% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{z22_metadata}
\alias{z22_metadata}
\title{Metadata}
\usage{
z22_metadata(name, type, area = c("all", "public", "user"))
}
\arguments{
\item{name}{Name of the variable, value, table, or statistic for which to
retrieve metadata.}

\item{type}{Whether \code{name} refers to a variable, value, table, or
statistic.}

\item{area}{If \code{type} is \code{table}, specifies the permission area
in which a table is stored. Can be \code{"all"}, \code{"public"}, or
\code{"user"}.}
}
\value{
An object of class \code{z22_metadata}, which is a list of relevant
information about an object.
}
\description{
Returns metadata on variables, values, tables, or statistics.
}
