% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{z22_statistics}
\alias{z22_statistics}
\title{Statistics}
\usage{
z22_statistics(
  selection = NULL,
  searchcriterion = c("code", "content"),
  sortcriterion = c("code", "content"),
  pagelength = 100L
)
}
\arguments{
\item{selection}{A search string with a maximum of 15 characters. The
statistics catalog is searched for the selection string. The
\code{searchcriterion} can be used to specify which column should be
searched. Search strings can use the wildcard operator \code{*}.}

\item{searchcriterion}{Column to search for the search string in
\code{selection}. Must be one of \code{"code"} or \code{"content"}.}

\item{sortcriterion}{Column by which to sort the result.
Must be one of \code{"code"} or \code{"content"}.}

\item{pagelength}{Number of results to return up to a maximum of 2,500.
Defaults to 100.}
}
\value{
A dataframe containing the code, description, and cubes of
all queried statistics.
}
\description{
Returns a list of statistics according to specified parameters.
}
\examples{
\dontrun{
# return all statistics
z22_statistics()

# sort by content name
z22_statistics(sortcriterion = "content")

# search for all codes starting with 3000
z22_statistics("3000*")
}
}
