% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{z22_tables}
\alias{z22_tables}
\title{Tables}
\usage{
z22_tables(
  statistic = NULL,
  variable = NULL,
  selection = NULL,
  area = c("all", "public", "user"),
  searchcriterion = c("code", "contents"),
  sortcriterion = c("code", "contents"),
  pagelength = 100L
)
}
\arguments{
\item{statistic}{Name of a statistic for which to return tables.}

\item{variable}{Name of a variable for which to return tables.}

\item{selection}{A search string with a maximum of 15 characters. The
statistics catalog is searched for the selection string. The
\code{searchcriterion} can be used to specify which column should be
searched. Search strings can use the wildcard operator \code{*}.}

\item{area}{If \code{type} is \code{table}, specifies the permission area
in which a table is stored. Can be \code{"all"}, \code{"public"}, or
\code{"user"}.}

\item{searchcriterion}{Column to search for the search string in
\code{selection}. Must be one of \code{"code"} or \code{"content"}.}

\item{sortcriterion}{Column by which to sort the result.
Must be one of \code{"code"} or \code{"content"}.}

\item{pagelength}{Number of results to return up to a maximum of 2,500.
Defaults to 100.}
}
\value{
A dataframe containing the code, description, and cubes of
all queried statistics.
}
\description{
Returns a list of statistics according to specified parameters.
}
\examples{
\dontrun{
# return all tables
z22_tables()

# sort by content name
z22_statistics(sortcriterion = "content")

# search for all codes starting with 1000A-100
z22_tables(selection = "1000A-100*")
}
}
