% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{z22_variables}
\alias{z22_variables}
\alias{z22_variables_from_statistic}
\title{Variables}
\usage{
z22_variables(
  selection = NULL,
  searchcriterion = c("code", "content"),
  sortcriterion = c("code", "content"),
  type = "all",
  pagelength = 100L
)

z22_variables_from_statistic(
  name,
  selection = NULL,
  searchcriterion = c("code", "content"),
  sortcriterion = c("code", "content"),
  type = "all",
  pagelength = 100L
)
}
\arguments{
\item{selection}{A search string with a maximum of 6 characters. The
statistics catalog is searched for the selection string. The
\code{searchcriterion} can be used to specify which column should be
searched. Search strings can use the wildcard operator \code{*}.}

\item{type}{Type of variable to include. This argument accepts several
multi-language values: \code{"all"} or \code{"alle"} includes all types of
variables. \code{"insgesamt"} includes variables about regional
totals. \code{"räumlich"} includes variables about spatial units.
\code{"sachlich"} or \code{"subject"} includes variables about subjects.
\code{"wert"} or \code{"value"} includes variables about specific values.
\code{"classifying"} or \code{"klassifizierend"} includes variables of
types subject, regional, and regional total. \code{"zeitidentifizierend"}
includes variables about temporal units. \code{"classifying"} or
\code{"klassifizierend"} includes classifying variable types "subject",
"regional", and "regional total".}

\item{pagelength}{Number of results to return up to a maximum of 2,500.
Defaults to 100.}

\item{name}{For \code{z22_variables_from_statistic}, a code of a statistic
whose variables to return. For \code{z22_variable}, the code of a variable
for which to return metadata.}
}
\value{
\code{z22_variables} and \code{z22_variables_from_statistic} return
a dataframe containing the code, content, type, and number of
values of a variable.

\code{z22_variable} returns a list of metadata with class \code{z22_details}
on a specific variable.
}
\description{
Return a list of variables according to the specified parameters.
}
\examples{
\dontrun{
# return the first 100 variables
z22_variables()

# search for age-related code
z22_variables("ALT*")

# search for age-related descriptions
z22_variables("Age*", searchcriterion = "content")

# return only space-related variables
z22_variables(type = "räumlich")

# return all variables from a specific statistic
z22_variables_from_statistic("1000A")

# return metadata on a variable
z22_variable("GEBEN1")
}
}
