% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z22-package.R
\docType{package}
\name{z22-package}
\alias{z22}
\alias{z22-package}
\title{z22: Official Gridded Data from the German Census 2022}
\description{
Quickly and efficiently retrieve gridded census data from Germany's 2011 and 2022 censuses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jslth/z22/}
  \item \url{https://jslth.github.io/z22/}
  \item Report bugs at \url{https://github.com/jslth/z22/issues}
}

}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org} [copyright holder]

}
\keyword{internal}
