% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z22-package.R
\docType{package}
\name{z22-package}
\alias{z22}
\alias{z22-package}
\title{\code{z22} uses a number of convenience options, including:}
\description{
\itemize{
 \item{\code{z22_kvp}: A logical value specifying whether to send POST requests
 or Key-Value Pair (KVP) requests.}
 \item{\code{z22_echo}: A logical value specifying whether requests should be
 echoed before sending. This can be useful to debug or to get to know
 what z22 does behind the scenes.}
 \item{\code{z22_language}: A character string that specifies the language
 used in outputs. Corresponds to the \code{language} parameter in the
 Zensus documentation. Can be either \code{en} or \code{de}. Defaults to
 \code{en}.}
}
}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org} [copyright holder]

}
\keyword{internal}
