% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z22-package.R
\docType{package}
\name{z22-package}
\alias{z22}
\alias{z22-package}
\title{z22: Official Gridded Data from the German Census 2022}
\description{
Provides fast and easy access to German census grid data from the 2011 and 2022 censuses \url{https://www.zensus2022.de/}, including a wide range of socio-economic indicators at multiple spatial resolutions (100m, 1km, 10km). Enables efficient download, processing, and analysis of large census datasets covering population, households, families, dwellings, and buildings. Harmonized data structures allow direct comparison with the 2011 census, supporting temporal and spatial analyses. Facilitates conversion of data into common formats for spatial analysis and mapping ('terra', 'sf', 'ggplot2').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jslth/z22/}
  \item \url{https://jslth.github.io/z22/}
  \item Report bugs at \url{https://github.com/jslth/z22/issues}
}

}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org} (\href{https://orcid.org/0000-0002-3451-3176}{ORCID}) [copyright holder]

}
\keyword{internal}
