% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{z22_decode_feat}
\alias{z22_decode_feat}
\alias{z22_decode_cat}
\alias{z22_translate_feat}
\alias{z22_translate_cat}
\title{Decode and translate features and categories}
\usage{
z22_decode_feat(features, lang = c("english", "german"))

z22_decode_cat(
  features,
  codes = NULL,
  lang = c("english", "german"),
  parallel = FALSE
)

z22_translate_feat(
  features,
  type = c("desc", "name"),
  lang = c("english", "german")
)

z22_translate_cat(categories, lang = c("english", "german"))
}
\arguments{
\item{features}{A character vector containing English or German feature
names. For \code{z22_translate_feat} and if \code{type} is \code{"desc"},
can also be the description associated with a feature.}

\item{lang}{Specifies the language of the output description. Can be
either \code{"english"} (default) or \code{"german"}. Note that the
English descriptions are only ad-hoc translations based off the German
originals.}

\item{codes}{A numeric vector containing category codes. If \code{NULL},
returns all category descriptions within the feature specified in the
\code{feature} argument.}

\item{parallel}{If \code{TRUE}, decodes categories in parallel, similar
to how \code{\link{mapply}} works. If \code{TRUE}, \code{features}
and \code{codes} must always have the same length and a dataframe with the
same number of rows is returned. If \code{FALSE}, \code{features} can have
a different length than \code{codes} and the output can be longer, depending
on how many more codes there are than features. Defaults to \code{FALSE}.
See examples.}

\item{type}{Whether to translate the feature name or description.}

\item{categories}{A character vector containing the English or German
descriptions of category codes.}
}
\value{
\code{z22_decode_feat} returns a character vector with the same
length as \code{features} containing the English or German descriptions
of the features or categories.

\code{z22_decode_cat} always returns a tibble with three columns
\code{feature}, \code{code}, and \code{label} where the label contains the
decoded description. If \code{parallel = TRUE}, the tibble contains the
same number of rows as the length of the input.
}
\description{
Helper functions to get the English or German descriptions associated with
a feature name or category code.

\itemize{
 \item{The \code{z22_decode_*} functions help to retrieve human-readable
 descriptions from abbreviated feature names}
 \item{The \code{z22_translate_*} functions help to translate feature names
 or descriptions from German to English or vice versa}
}
}
\examples{
# Get english description of feature names
z22_decode_feat(c("ALTER_KURZ", "ALTER_10JG"))

# You can use both german and english feature names.
# English feature names are not official but might be better
# for code readability.
z22_decode_feat(c("age_short", "age_long"))

# Get all categories for specific features
z22_decode_cat(c("build_size", "build_heating"))

# Translate feature names
z22_translate_feat(
  c("build_size", "build_heating"),
  lang = "german",
  type = "name"
)

# Translate category labels
z22_translate_cat("Eingetr. Lebenspartnerschaft")

# The `lang` argument only changes the output. The input can be any language
z22_decode_cat("foreigner_share", lang = "german")

# The parallel argument allows you to extract a single category label for
# each combination of feature and category code
z22_decode_cat(c("build_heating", "build_heating"), c(1, 2))
z22_decode_cat(c("build_heating", "build_heating"), c(1, 2), parallel = TRUE)
}
