% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes.R
\name{z22_grid}
\alias{z22_grid}
\title{Get INSPIRE grid}
\usage{
z22_grid(res, rasterize = FALSE, as_sf = FALSE)
}
\arguments{
\item{res}{Resolution of the grid dataset. Can be \code{"100m"}
or \code{"1km"}.}

\item{rasterize}{If \code{TRUE} and the \code{terra} package is installed,
converts the attribute coordinates to a \code{\link[terra:rast]{SpatRaster}}.}

\item{as_sf}{If \code{TRUE} and the \code{sf} package is installed,
converts the attribute coordinates to an \code{sf} tibble.}
}
\value{
A tibble, \code{\link[terra:rast]{SpatRaster}} or
\code{\link[sf::st_as_sf]{sf}} tibble depending on the \code{rasterize}
and \code{as_sf} arguments.
}
\description{
Retrieve the entire INSPIRE grid.

Unlike the attribute grids retrieved from \code{\link{z22_get_attribute}},
the INSPIRE grid encompasses the entire area of Germany. You can thus use
it to join with the incomplete attribute grids from \code{z22_get_attribute}
to create a complete dataset.
}
\examples{
\donttest{# Get high-res grid
z22_grid()

# Get low-res grid as raster
z22_grid("1km", rasterize = TRUE)}
}
